requirements.txt
streamlit>=1.32
pillow>=10.0
numpy>=1.24
scipy>=1.10
